Clean and refactoring of branch 'devGan' #2
Closes #2. Branch refactoring: all the elements have been moved to different modules following the already existent structure. In addition, the structure of the input functions is now more consistent and just requires calling generate_input_fn; 'batch_size' and 'noise_dim' are passed via 'params' now. Redundant and unused code cleaned.
