+++ src/core/core_model_estimator.py
@@ -3,275 +3,236 @@
 """
 
 import os
-from absl import logging
+import numpy as np
 
 import tensorflow as tf
+USE_ALTERNATIVE = False
+try:
+    from tensorflow.data import Dataset
+except ImportError:
+    tf.logging.warning('Using alternative settings due to old TF version')
+    Dataset = tf.data.Dataset
+    USE_ALTERNATIVE = True
 from tensorflow.contrib import tpu
+from tensorflow.contrib.cluster_resolver import TPUClusterResolver #pylint: disable=E0611
+from tensorflow.python.estimator import estimator
 
-from .core_datamanager_estimator import DataManagerTPU as DataManager
+from util.image_postprocessing import convert_array_to_image
+
+tfgan = tf.contrib.gan
+queues = tf.contrib.slim.queues
+layers = tf.contrib.layers
+ds = tf.contrib.distributions
+framework = tf.contrib.framework
+
 
 class CoreModelTPU(object):
 
     def __init__(self,
-                 tf_session: tf.Session,
-                 learning_rate: float,
-                 training_dataset: DataManager = None,
-                 validation_dataset: DataManager = None,
-                 output_path: str = '../outputs',
+                 model_dir: str,
+                 data_dir: str,
+                 learning_rate: float = 0.0002,
+                 noise_dim: int = 64,
+                 batch_size: int = 128,
+                 iterations_per_loop: int = 100,
+                 num_viz_images: int = 100,
+                 eval_loss: bool = False,
+                 train_steps_per_eval: int = 100,
+                 num_eval_images: int = 100,
                  use_tpu: str = False,
-                 tpu_name: list = [],
-                 data_dir= '/mnt/iowa_bucket/cifar10/data/'
+                 tpu: str = '',
+                 tpu_zone: str = None,
+                 gcp_project: str = None,
+                 num_shards: int = None
                 ):
+        """Wrapper class for the model.
+
+        Args:
+            model_dir (str): Model directory
+            data_dir (str): Data directory
+            learning_rate (float, optional): Defaults to 0.0002.
+            batch_size (int, optional): Defaults to 1024.
+            iterations_per_loop (int, optional): Defaults to 500. Iteratios per loop on the estimator.
+            num_viz_images (int, optional): Defaults to 100. Number of example images generated.
+            eval_loss (bool, optional): Defaults to False.
+            train_steps_per_eval (int, optional): Defaults to 100.
+            num_eval_images (int, optional): Defaults to 100. Number of eval samples.
+            use_tpu (str, optional): Defaults to False.
+            tpu (str, optional): Defaults to ''. TPU to use.
+            tpu_zone (str, optional): Defaults to None.
+            gcp_project (str, optional): Defaults to None.
+            num_shards (int, optional): Defaults to None.
+        """
+
         self.data_dir = data_dir
-        if output_path[-1] == '/':
-            output_path = output_path[:-1]
-        self.output_path = output_path + '/' + self.__class__.__name__
-
-        self.session = tf_session
-
-        # TODO Get rid of the .datasource thing
-        self.dataset = {}
-        # if training_dataset: self.dataset['train'] = training_dataset.datasource
-        # if validation_dataset: self.dataset['validation'] = validation_dataset.datasource
-
-        self.datasource = {}
-        self.datasource['train'] = training_dataset
-        self.datasource['validation'] = validation_dataset
-
-        self._train_model =  True if training_dataset is not None else False
-        self._validate_model = True if validation_dataset is not None else False
-        self.learning_rate = learning_rate
+        if model_dir[-1] == '/':
+            model_dir = model_dir[:-1]
+        self.model_dir = model_dir + '/' + self.__class__.__name__
 
         self.use_tpu = use_tpu
+        self.tpu = tpu
+        self.tpu_zone = tpu_zone
+        self.gcp_project = gcp_project
+        self.num_shards = num_shards
 
+        self.learning_rate = learning_rate
+        self.noise_dim = noise_dim
+        self.batch_size = batch_size
+        self.iterations_per_loop = iterations_per_loop
 
-    def define_model(self, data_source: DataManager , mode: str): #pylint: disable=E0202
-        """Definition of the model to use. Do not modify the function here
-        placeholder for the actual definition in `model/` (see example)
+        self.num_viz_images = num_viz_images
+        self.eval_loss = eval_loss
+        self.train_steps_per_eval = train_steps_per_eval
+        self.num_eval_images = num_eval_images
 
-        Args:
-            data_source (DataManager): Data manager object for the input data
-            mode (str): Training and testing? # TODO Properly implement
+    def generate_model_fn(self):
+        """Definition of the model function to use. It should return a model_fn
+        that takes arguments: features, labels, mode and params; and returns a
+        set of TPUEstimatorSpec.
 
-        Raises:
-            NotImplementedError: Model has to be implemented yet (in a separate instance in model/)
+        Do not modify the function here this is a placeholder for the actual
+        definition in `model/` (see example)
         """
 
-        raise NotImplementedError('No model defined.')
+        raise NotImplementedError('No model function defined')
+
+    def make_config(self):
+        """Generates a config file for the tf.Estimators"""
+
+        tpu_cluster_resolver = None
+        if self.use_tpu:
+            tpu_cluster_resolver = tf.contrib.cluster_resolver.TPUClusterResolver(
+                self.tpu,
+                zone=self.tpu_zone,
+                project=self.gcp_project)
+
+        config = tpu.RunConfig(
+            cluster=tpu_cluster_resolver,
+            model_dir=self.model_dir,
+            tpu_config=tf.contrib.tpu.TPUConfig(
+                num_shards=self.num_shards,
+                iterations_per_loop=self.iterations_per_loop))
+
+        params = {
+            'data_dir': self.data_dir,
+            'noise_dim': self.noise_dim
+            }
+        return config, params
 
     def build_model(self):
-        """ Build the model. """
-        if self.use_tpu:
-            self._tpu_build()
-        else:
-            self._regular_build()
+        """Builds the tensorflow model"""
 
-    def _tpu_build(self):
-        """Build with TPUEstimators for TPU usage"""
-        def _define_model(features, labels, mode, params):
-            data_source = (features, labels)
-            self.outputs = {}
-            self.losses = {}
-            self.otters = {}
-            outputs, losses, others = self.define_model(data_source, mode)
+        model_fn = self.generate_model_fn()
 
-            if mode == tf.estimator.ModeKeys.EVAL:
-                return tpu.TPUEstimatorSpec(
-                    mode=mode, loss=losses, eval_metrics=others)
-            if mode == tf.estimator.ModeKeys.PREDICT:
-                return tpu.TPUEstimatorSpec(
-                    mode=mode, predictions=outputs
-                )
-            if mode == tf.estimator.ModeKeys.TRAIN:
-                self.losses['train'] = losses
-                self._build_optimizer(tpu_support=True)
-                if not len(self.optimize_ops) == 1:
-                    logging.error('Implementati Error: More than one optimizer defined')
-                    logging.warning(' [*] Selecting only the first optimizer')
-                return tpu.TPUEstimatorSpec(
-                    mode=mode, loss=losses[0], train_op=self.optimize_ops[0]
-                )
+        tf.logging.info('Start')
+        config, params = self.make_config()
 
-        tpu_name = ['node-1'] # TODO Bring outside
-        tpu_iterations = 500  # TODO Bring outside
-        tpu_cluster_resolver = tf.contrib.cluster_resolver.TPUClusterResolver(
-            tpu_name)
+        # TPU-based estimator used for TRAIN and EVAL
+        self.est = tf.contrib.tpu.TPUEstimator(
+            model_fn=model_fn,
+            use_tpu=self.use_tpu,
+            config=config,
+            params=params,
+            train_batch_size=self.batch_size,
+            eval_batch_size=self.batch_size)
 
-        run_config = tf.contrib.tpu.RunConfig(
-            model_dir=self.output_path,
-            cluster=tpu_cluster_resolver,
-            session_config=tf.ConfigProto(
-                allow_soft_placement=True, log_device_placement=True),
-            tpu_config=tpu.TPUConfig(tpu_iterations),
-            )
-
-        self.estimator = tpu.TPUEstimator(
-            model_fn=_define_model,
-            use_tpu=True,
-            train_batch_size=32*4, #self.dataset['train'].batch_size,
-            eval_batch_size=32*4, #self.dataset['validation'].batch_size,
-            config=run_config,
-            params={"data_dir": self.data_dir}
-        )
+        # CPU-based estimator used for PREDICT (generating images)
+        self.cpu_est = tf.contrib.tpu.TPUEstimator(
+            model_fn=model_fn,
+            use_tpu=False,
+            config=config,
+            params=params,
+            predict_batch_size=self.num_viz_images)
 
 
-    def _regular_build(self):
-        """Normal build for CPU/GPU usage"""
-        # This overwrites define_model, is that ok?
-        self.define_model = tf.make_template(self.define_model.__name__,  #pylint: disable=E1101
-                                             self.define_model,
-                                             create_scope_now_=True)
-
-        self.outputs = {}
-        self.losses = {}
-        self.otters = {}
-
-        def _build(mode):
-            outputs, losses, others = self.define_model(data_source=self.dataset[mode], mode=mode)
-            self.outputs[mode] = outputs
-            self.losses[mode] = losses
-            self.otters[mode] = others
-            if mode == 'train':
-                self._build_optimizer()
-
-        # TODO Move clean and summary to proper section
-        self.summary_ops = {}
-        if self._train_model:
-            _build('train')
-            summary = []
-            for idx, loss in enumerate(self.losses['train']):
-                summary.append(
-                    tf.summary.scalar(name='train/loss_{}'.format(idx), tensor=loss))
-            for idx, element in enumerate(self.otters['train']):
-                summary.append(
-                    tf.summary.scalar(name='train/otter_{}'.format(idx), tensor=element))
-            self.summary_ops['train'] = tf.summary.merge(summary)
-
-        if self._validate_model:
-            _build('validation')
-            summary = []
-            for idx, loss in enumerate(self.losses['validation']):
-                summary.append(
-                    tf.summary.scalar(name='val/loss_{}'.format(idx), tensor=loss))
-            for idx, element in enumerate(self.otters['validation']):
-                summary.append(
-                    tf.summary.scalar(name='val/otter_{}'.format(idx), tensor=element))
-            self.summary_ops['validation'] = tf.summary.merge(summary)
-
-        self.writer = tf.summary.FileWriter(self.output_path,
-                                            self.session.graph)
-        self.saver = tf.train.Saver()
-        # TODO Add routine to save
-        logging.info('Model construction complete.')
-
-    def _build_optimizer(self, optimizer_to_use=tf.train.AdamOptimizer, tpu_support=False):
-        """Buids the optimizer(s) to minimize the loss(es) of the model.
+    def train(self,
+              train_steps,
+              generate_input_fn):
+        """Train the model
 
         Args:
-            optimizer_to_use (tf optimizer, optional): Defaults to tf.train.AdamOptimizer. Which
-                optimizer to use.
-            tpu_support (bool, optional): Defaults to False. If the optimizer has to support shard
-                optimier, required for TPU usage.
+            train_steps (int): Numer of training steps
+            generate_input_fn (function): Function that resturns input_fn
+            function. (see example or tf.Estimator documentation)
+            noise_input_fn (function): input_fn that returns a noise vector
         """
-        self.optimize_ops = []
-        for loss in self.losses['train']:  # TODO Create apropoiate external training scheme
-            optimize_op = optimizer_to_use(
-                learning_rate=self.learning_rate
-            )
-            if tpu_support:
-                optimize_op = tpu.CrossShardOptimizer(optimize_op)
-            optimize_op = optimize_op.minimize(
-                loss=loss,
-                global_step=tf.train.get_global_step()
-            )
-            self.optimize_ops.append(optimize_op)
-        logging.info('Optimizers built')
 
-    def train(self, steps, input_fn=None):
-        if self.use_tpu:
-            self._tpu_train(steps, input_fn)
-        else:
-            self._regular_train(steps)
+        current_step = estimator._load_global_step_from_checkpoint_dir(self.model_dir)   # pylint: disable=protected-access,line-too-long
+        tf.logging.info('Starting training for %d steps, current step: %d' %
+                        (train_steps, current_step))
+        tf.gfile.MakeDirs(os.path.join(self.model_dir, 'generated_images'))
 
-    def _tpu_train(self, steps, input_fn):
+        while current_step < train_steps:
+            next_checkpoint = int(min(current_step + self.train_steps_per_eval,
+                                    train_steps))
+            tf.logging.info('Step: %s  -- (Next checkpoint %s)', current_step, next_checkpoint)
+            self.est.train(input_fn=generate_input_fn('TRAIN'),
+                        max_steps=next_checkpoint)
+            current_step = next_checkpoint
+            tf.logging.info('Finished training step %d' % current_step)
 
-        # def _input_fn(params):
-        #     featuers, labels = self.datasource['train'].input_fn(params['batch_size'])
-        #     return featuers, labels
+            if self.eval_loss:
+                # Evaluate loss on test set
+                metrics = self.est.evaluate(input_fn=generate_input_fn('EVAL'),
+                                        steps=max(self.num_eval_images // self.batch_size,1))
+                tf.logging.info('Finished evaluating')
+                tf.logging.info(metrics)
 
-        self.estimator.train(
-            input_fn=input_fn,
-            max_steps=steps)
-        logging.info('Es ist train?')
-        self.estimator.evaluate(
-            input_fn=self.dataset['validation'],
-            steps=steps/50
-        )
-        print("\nTest set accuracy: {accuracy:0.3f}\n".format(**eval_result))
+            # Render some generated images
+            generated_iter = self.cpu_est.predict(input_fn=generate_input_fn('PREDICT'))
+            images = [p['generated_images'][:, :, :] for p in generated_iter]
+            if len(images) != self.num_viz_images :
+                tf.logging.info('Made %s images (when it should have been %s',
+                    len(images), self.num_viz_images )
+                images = images[:self.num_viz_images ]
 
-    def _regular_train(self, steps):
-        # Initialize or check if checkpoint # TODO add checkpoint manager
-        self.session.run(tf.global_variables_initializer())
-        initial_step = self._restore()
+            # assert len(images) == self.num_viz_images
+            image_rows = [np.concatenate(images[i:i+10], axis=0)
+                            for i in range(0, self.num_viz_images , 10)]
+            tiled_image = np.concatenate(image_rows, axis=1)
 
-        fetches = {}
-        fetches['optimize_ops'] = self.optimize_ops
-        # fetches['losses'] = self.losses['train']
-        # if self.otters['train']:
-        #     fetches['others'] = self.otters['train']
-        fetches['summary_ops'] = self.summary_ops['train']
+            img = convert_array_to_image(tiled_image)
 
-        for step in range(initial_step, steps):  # TODO start from checkpoint steps
-            # TODO clean code and optimize ops
-            train_out = self.session.run(fetches=fetches)
-            self.writer.add_summary(train_out['summary_ops'], global_step=step)
-            if step % 50 == 0: # TODO every how many steps? Automate?
-                val = self._validate(step)
-                logging.info('Step {} -- Validation result: {}'.format(step, val))
-            if step % 1000 == 0:  # For now just another arbitrary number (how heavy is saving?)
-                self._save(step)
-        logging.info('Done training.')
+            step_string = str(current_step).zfill(5)
+            file_obj = tf.gfile.Open(
+                os.path.join(self.model_dir,
+                                'generated_images', 'gen_%s.png' % (step_string)), 'w')
+            img.save(file_obj, format='png')
+            tf.logging.info('Finished generating images')
 
-    def _validate(self, global_step):
-        """ Run network on validation set """
-        # Todo clean summaries and add example outputs
-        fetches = {}
-        fetches['losses'] = self.losses['validation']
-        if self.otters['train']:
-            fetches['others'] = self.otters['validation']
-        fetches['summary_ops'] = self.summary_ops['validation']
-        validation_out = self.session.run(fetches=fetches)
-        self.writer.add_summary(validation_out['summary_ops'], global_step=global_step)
-        del validation_out['summary_ops']
-        return validation_out
 
-    def _save(self, step):
-        """Save the model weights.
+    def save_samples_from_data(self, generate_input_fn):
+        """Sample some images from the data and print them as images.
 
         Args:
-            step (int): Training step.
+            input_fn (function): Input function used on the model.
         """
 
-        output_path = self.output_path + '/checkpoints/'
-        if not os.path.isdir(output_path):
-            os.makedirs(output_path)
-        self.saver.save(self.session, save_path=output_path,global_step=step)
+        tf.gfile.MakeDirs(os.path.join(self.model_dir, 'generated_images'))
 
-    def _restore(self):
-        """Restore the trained variables from the last stored checkpoint
+        config, params = self.make_config()
+        # Get some sample images
+        # CPU-based estimator used for PREDICT (generating images)
+        data_sampler = tf.contrib.tpu.TPUEstimator(
+            model_fn=lambda features, labels, mode, params: tpu.TPUEstimatorSpec(mode=mode, predictions=features['real_images']),
+            use_tpu=False,
+            config=config,
+            params=params,
+            predict_batch_size=self.num_viz_images )
 
-        Returns:
-            int: The training step when this model was saved.
-        """
+        sample_images = data_sampler.predict(input_fn=generate_input_fn('TRAIN'))
+        tf.logging.info('That ran')
+        images = []
+        for i in range(self.num_viz_images ):
+            images.append(next(sample_images))
 
-        output_path = self.output_path + '/checkpoints/'
-        checkpoint = tf.train.latest_checkpoint(output_path)
-        if checkpoint:
-            self.saver.restore(self.session, save_path=checkpoint)
-            restored_step = int(checkpoint.split('-')[-1])  # Robust enough?
-            return restored_step
-        logging.info('Starting training from scratch.')
-        return 0
+        image_rows = [np.concatenate(images[i:i+10], axis=0)
+                    for i in range(0, self.num_viz_images , 10)]
+        tiled_image = np.concatenate(image_rows, axis=1)
+        img = convert_array_to_image(tiled_image)
 
-    def evaluate(self):
-        pass
+        file_obj = tf.gfile.Open(
+            os.path.join(self.model_dir,
+                            'generated_images', 'sampled_data.png'), 'w')
+        img.save(file_obj, format='png')
+        tf.logging.info('File with sample images created.')
