+++ src/datamanager/__init__.py
@@ -1,9 +1,9 @@
 """Set of core packages"""
 
 from .CIFAR_manager import CIFAR10
-from .cq500_manager import CQ500
+from .CIFAR_input_functions import generate_input_fn
 
 __all__ = (
     'CIFAR10',
-    'CQ500'
+    'generate_input_fn'
 )
